Added lineplot() functionality
Added lineplot() functionality, started work on basic_boxplot,
and made some changes to the DB.

Took out some of the html.parsing as it seemed to be causing issues.
Now that I'm doing SQL differently, it might be effect it, which is
why I retained the include statement.

I also changed some of the fstring formatting on the DB code,
as that is vulnerable.

Apparently, the way I'm doing my queries in fetch_all_months() and
fetch_all_years() is the way to mitigate SQL injection.

I changed a small line in scrape_weather.py to add in 0's to
single digit months.

Gitignore isn't working exactly correctly for me for __pycache__,
but I'm sure we will figure that out.