Fixed error with current date in LinePlot
When opening a LinePlot for the current month, the graph would throw errors
if the month wasn't fully complete
(ie: lineplot for March when the current date is the 23rd)

Added a check for null data, and it will replace with 0's if found.

Also changed the "start/end" year for the Boxplot, so that
the first year the user enters is the oldest, and the second is the
newest.